requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Cors==2.0.0
Flask-Migrate==1.3.1
Flask-RESTful==0.3.2
Flask-SQLAlchemy==2.0
Flask-Script==0.6.2
Flask-SeaSurf==0.1.22
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
Pillow==2.2.2
PyYAML==3.10
Pygments==1.6
SQLAlchemy==0.9.9
Sphinx==1.2
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.10.4
XlsxWriter==0.5.7
alembic==0.7.5.post2
amqp==1.4.6
aniso8601==0.92
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.18
celery==3.1.13
decorator==3.4.2
docutils==0.11
enum34==0.9.19
factory-boy==2.3.0
flower==0.6.0
ipdb==0.8
ipython==1.1.0
itsdangerous==0.24
kombu==3.0.21
log4mongo==1.3.0
natsort==3.0.2
nose==1.3.0
postgres==2.1.2
psycopg2==2.5.4
pycrypto==2.6.1
pymongo==2.6.3
python-dateutil==2.2
python-magic==0.4.6
pytz==2015.2
redis==2.8.0
requests==2.0.1
sh==1.09
simplejson==3.6.5
six==1.9.0
speaklater==1.3
sqlalchemy-migrate==0.9.5
sqlparse==0.1.14
suds-jurko==0.6
tornado==3.1.1
uWSGI==1.9.20
wheezy.captcha==0.1.29
